Fix logout button
A very long road to get to this simple fix. Log out did not work
because, for XSS security, it is now required to be a DELETE method
(or a POST-as-DELETE). Converting to a DELETE method did not work
because rails.js was not invoked. Rails.js was not invoked because
the latest round of upgrades caught us up in the migration from
Prototype to JQuery, and we no longer had a usable way to invoke
rails.js. So by this last round of giant commits to upgrade to Rails
3.1 and the Asset Pipeline, we've caught up all round and simple
fixes like this one now have an effect. Let there be progress.

Test #91