WIP porting out report framework
With this commit the regional change interpreters test clean.
Continental change interpreters, and continental and regional
range analyses, appear but do not test clean.

A missing query from AED Crystal, "ChangesInterpreter" has been
ported and tests clean, which provides the mappings for aed2007
data at the regional level.

The views used internally for the change interpreter scaling are
much terser in the work done here for the regions. Eventually
this work should be backported to the country interpreter; FIXME
notes are in the SQL view generator scripts for this.