Reorder SQL query for proper range aggregates
The issue with the regional and continental range aggregates was
just a bad SQL query order. We were LEFT JOINing subsequent
columns on the first column, which happened to be Known range;
thus surveys that only touched on Possible range were excluded.
This change reorders the query to start with Total range, then
LEFT JOIN in the columns for Known and Possible, which are always
subsets of Total. This tests green.

The same result is obtained whether using the historical SurvRang
or intersecting the polygons with PostGIS. At one point we had a
query which could regenerate SurvRang from PostGIS. Since then it
has been ported forward to the Rails data model, in which state it
remains untested - but its original test was to generate SurvRang
from the aed2007 data set, which it did with only small fractional
discrepancies from the historical ESRI answers.

Test #78 (again ... finally)