Add Country, Species, Range States to live schema
This diverges significantly from what was done in the aed* schemas
because we are introducing multi-species capability here. The
Country table created here is minimal compared to what is in aed*,
because it need not keep static copies of values computed in views.

Also, it is meant more as a link target than a data container. It
should be a systemwide ISO-compliant list. A future enhancement
would be to add a geometry column.

The structure provided here of the Species and Range State link
tables is representative of how this should be done.

Basic Rails CRUD could be done to manage these tables if there is
a reason to manage them on the web instead of PostgreSQL directly.
I've created the tables with the appropriate artifacts (timestamps,
numeric IDs) and populated them using ActiveRecord operations
instead of raw SQL, to allow this evolution in future if desired.

See #31
See #123