WIP better flow from initial submission
The initial submission short-circuits you to the associated
population submission (if you chose to submit population data).
This general pattern should be repeated throughout the forms.
Create a partial in app/views/application that shows a prettified
restatement of the information in the associated model. Add
the prettifying styles in the associated SCSS stylesheet in
app/assets. Reference the partial in downstream views.