Functional editing using new UI through all AS
It should now be possible to create an Aerial Sample survey using
the new proposed UI for assembling a survey. There is also now
a nonfunctional 'attach a file' link to show how the attachments
could be integrated into that pane.

The patterns in the AS strata can be made DRY so they port easily
to other survey types.

This does more clearly expose the problem of where to store
(or calculate) aggregate estimates when reporting by stratum.
Some discussion will be needed to sort that aspect out properly.