Implement list of submissions
Several things going on here. The Slim gem is added
(http://slim-lang.org) which provides a generation more advanced
templating than HAML. It is very similar, but faster and more
readable, and more portable to Jade (JS) and other modern
template languages. The population_submissions index view is now
coded in Slim. It is OK to have a mix of HAML and Slim, but as we
touch HAML files it would be nice to convert them to Slim.

The design of this calls for showing a single estimate for surveys,
which we have never really implemented before now in the new model.
There is a Ruby function added to the population_submission model
now which iterates over the counts and strata enclosed and adds
the estimate column. That is probably naive, but yields a result
sufficient for demo to the donor. FIXMEs in the code reflect that
this should be more properly moved to PostgreSQL as a view, and
any pooling required across strata be implemented.

Test #118