Encoding/escaping issue
Côte d'Ivoire is an interesting case because it contains both a
non-ASCII-7bit character (ô) and a single quote. The single quote
here was actually the problem; the country name is resolved
to a code in the database using a direct SQL query, in order to
be able to access the aed[year] schema. The query was not escaped,
allowing the single quote to cause a syntax error in PG. The PG
error message incorporates the offending string, which has the ô
character in it. Somewhere up the chain of the pg gem and/or
ActiveRecord, this untyped (ASCII-8BIT) error message is operated
on for display, and the presence of the ô produces a character set
mismatch error instead of the actual syntax error.

Test #85