Introduce flip left/right in forms; delete enabled
Known issues: "Are You Sure" confirmation fires twice. The right
side of the page is meant to have a pullout indicator showing which
part of the summary on the left you are working with. The delete
button is too small and too close to the edit button; probably we
want to move delete to the form so you have to edit something,
THEN click the link in the form area to delete it. But, this works
and bad data can be removed.

Removing a submission and orphaning its population submissions
will result in some inaccessible data.

Test #144 (just that delete works, not the formatting issues)