Replace iframe/target=_top hack with MapsV3 API
This is a long overdue commit to the mapping engine that allows
Rails to draw the popups over Fusion Tables. There are two key
benefits here: first, we only need to send geometries and IDs to
FT, reducing the chance of inadvertent data exposure (even though
we tick all the appropriate limitation boxes in FT). Second,
this approach draws the popups in the context of our app's
stylesheets and javascript scope. So we have a lot more stylistic
and operational control over what appears in them.