Point to fusion table with unioned polygons
This allows the entire range map to be displayed at all zooms.
Also, the information exported in this FT is minimal (just the
geometry and a 1 or 0 indicating the range quality).

To get this map, we use ST_Union followed by ST_Dump - this
yields about 500 polygons, none of which should exceed FT
complexity limits.

See #134