Display a message when no estimates
It hurts performance to look-ahead in the nav and see whether
there are actually any estimate in the thing being linked to.
Can we live with this - if you click to a country where there
are no estimates, it prints a message to that effect where the
table would normally be.

Test #102