Armor against inappropriate editing
This provides controller-level protection against users making
unauthorized edits. Next step is to update the UI to take away
the edit/delete links in the first place. For this, the
boolean edit_allowed method can be used to handle all the logic
of whether or not the operation will be allowed in context.

See #120