"Current" reference database source exposed
This creates the links to show the references exposed from the
aed2007 / aedcurrent references table. To produce the AESRs, these
went additional steps through reference management and desktop
publishing software. So there is some loss involved and the table
does not contain an accurate image of what was published, although
it is a good representative start.

The references page (/report/references) shows the raw database
source of each citation so you can see all the fields. It also
shows a formatted version of the reference based on some rudimentary
formatting code in this checkin.

If we think the original reference table can be brought up to speed
somehow with all the references that should be shown, then we can
undertake the remainder of the effort to make the formatter 100%
compliant with the IUCN standard + AESR extensions, e.g. personal
communications and questionnaire replies.

My guess is that the reference table could be cleaned, either on
the web using some basic Rails CRUD scaffolding, or with a copy
of the table in Excel or Access, given a day of dogged
determination and a bottle of ibuprofen ... so I think this is
the best way forward. I think the formatter, even though not 100%
perfect, shows a lot of promise for something we created quickly.

Marking Test and pushing onto staging for internal review and
comment, but not ready to close this quite yet!

Test #23