Unknowns in IR
First: Add "unknown" for frequency of re-sighting

This was possible because this is a text field and we can add this
option explicitly.

Second: Remove validation requirement for year population started

This is an integer field, so can only contain a number. "Unknown"
is not a valid value for the underlying database column. To achieve
a similar effect, I removed the validation rule so that this field
can be omitted, yielding a NULL in the database and therefore no
answer.

Test #223