Fix model issues with countries, species
The relationships for countries and species were not properly
modelled in Rails, in part because these used to be text fields
that were later converted to containing numeric data. The
associated migration may fail if you have text data in
country_id or species_id; clean it out if it is present. Backup
recommended before running the associated migration.

This impacts country sort behavior.

Test #234 (some more)