"Move" MIKE Site fields to the stratum level
This commit adds the fields to each stratum and changes the UI
and model to support this. It does not REMOVE the fields on the
submission object nor does it attempt to copy data residing
there. That should be done on the database directly. When all
data in the field is verified migrated, we can remove the field
from the submissions table.

Test #281