Support for preview mapping of input zones
Assuming that there is a multipolygon geometry stored with a given
input zone, this will throw a quick preview of the convex hull of
that geometry onto a Google Map on the summary page. It is not as
high quality (e.g. nested shapes, etc.) as one gets with the Fusion
Tables version, but the advantage is that it is drawn in realtime
from PostGIS and can quickly be used to spot where the zone is
and whether the geometry makes any sense.