Handle upper range case for IR
Since we can't really detect "IR-ness" here, what we are looking
for is a D with an estimate and an upper range, but no lower
range. If it had a lower range, the existing averaging logic
would catch it and asterisk it. This passes for the Chinko Basin
example and has a reasonably low risk of other side effects.

Test #304