Use bootstrapped layout for homepage
In addition, I simplified away the "ticker" features around the
3 most recent survey locations on Fusion Tables/Google Maps. This
needs to be rewritten for Leaflet if we want to keep it. What's
left here is just the basics of the page. Obviously it needs a
little redesign too -- it's not actually that responsive since
everything is wrapped in a .col-xs-12. But the technical
underpinnings are here now.