Wash and DRY views after toolchain upgrade.
Following a toolchain upgrade, views have been reviewed,
corrected, and, when possible, DRYed up, in order to make
the site functional again.

Note that some views need polish in terms of design, but all
should be functional and code should be in a state where we
aren't cleaning the same thing multiple times.

Major notes:

* Fixed login issue with new Devise
* Devise views updated to use @token for mailers
* Main layouts updated to use proper container/row/col format
* Reporting views cleaned up, split into partials
* New breadcrumbs view for reports, preview reports
* Removed excess views from survey/count CRUD, use single
  layout whenever possible
* Moved some repeated text into en.yml, particularly footnotes