Workaround a regression
In the view refactor the population_submission show page regressed.
So, e.g.

http://www.elephantdatabase.org/population_submissions/440

on the maintenance branch, shows some text about the population
submission (but no map), but

http://dev.elephantdatabase.org/population_submissions/440

on the master branch crashes trying to show a map. This commit
works around the crash but I could not very easily tell what had
gone wrong other than that the show method in survey_crud was
updated for more DRY views and this was probably biting us here.