Unzipping and shapefile reading bits
Using rgeo, which is a PROJ/GEOS wrapper for Ruby, instead of
invoking external shell commands. rgeo also underlies the postgis
adapter so this should give us economical and maintainable syntax
for moving this data into ActiveRecord models.

Also included rgeo-geojson to simplify the next step of showing
the acquired geometries on a Leaflet map.