New viewer for stratum data
This is built on the rgeo, leaflet, geojson stack and discards the
old Google Maps function with hand-built polygons. Accordingly,
this now correctly previews multipolygons which clears the
longstanding issue #331.

rgeo has trouble with a geometry column named 'geometry' due to a
name conflict; also the existing survey_geometry.geometry column
had a mix of 2d/3d and different SRID data in it. A migration here
creates a new 'geom' column but doesn't do anything to populate or
use it higher in the stack.

For now, a manual query is necessary to migrate old data into this
column:

update survey_geometries set geom=ST_Force2D(geometry);

This will be added to a migration later that is capable of also
updating the rest of the view stack to track this change.

Fix #331