Repair submission step
There was a regression around the submission page, the last step
in finalizing a survey. This page contained the abstract and
citation entry fields, as well as the toggles for "submitted" and
"released" for superusers. Getting this page to display again
makes things work largely as before, although we may want to
revisit the value of the submission step as currently implemented.

Test #394