Massive migration for trivial gain, exposes issues
I needed to make a minor migration to the analyses table so that
it could be managed more easily by Rails. Along the way I wanted
to fix the typing of the analysis_name column, which was doing an
inefficient text-to-string join in a number of places in the view
stack.

You can't change the type of a column used in an active view, so
making this trivial change requires rebuilding the view stack.
Should be no problem, just re-run the ETL scripts for this.

However, it simply wouldn't work for a host of reasons, which have
now been corrected, but may shed light on other heisenbugs we
are chasing in the calculations.

As written (by me - nobody else to blame), the ETL scripts would
work to update the views, but only if run on top of an existing
view stack. Due to some missing "if exists" clauses and so on, the
scripts would fail if run on an empty database. Added the missing
clauses to fix this.

The scripts weren't compatible with the SqlReader Carl added to
facilitate running the ETL scripts under migration control, due
to their comment syntax. Changed to /* inline comment */ style to
fix this.

Finally, the worrisome part, a small set of these views had been
since converted to static tables. Some of these were done for
performance reasons long ago -- the "recalc" feature impacts
these

      dpps_sums_region_category_reason

But this view was also converted to a static table

      i_dpps_sums_country_category_reason

Which I didn't do, and was polluting the system with some stale
data. In fact, "recalc" was literally just copying this stale
data into the static table dpps_sums_country_category_reason.

This migration drops the static version of that table and resets
it to dynamic calculation.