2007 geometry recapture script
This is implemented as a migration, but to get a useful result
from the migration, you must first have a copy of an original
AED 2007 shapefile loaded at "julian_2007" -- load the >3Mb
dumpfile or a shapefile directly to obtain this, or just get
a database image that has already had this migration run.

The migration emits a (hopefully) useful log of what it
recaptures.