Country level mapping
This shows a version of the Replacement Manager map on a tab in
the Country reports page. This is a largish commit because it
adds two new bits of plumbing: the ability to deferred load a
map on demand (as opposed to when the page is created) and also
some model mechanics for selectively including just those strata
included in a given analysis. These bits of plumbing should
be useful later in providing exports for printed report
generation. The model plumbing could be done with greater
efficiency by leveraging existing PostgreSQL views
if it needs to be accelerated -- with the tradeoff of tightly
binding the code to those views and needing to maintain both
SQL and Ruby code around this feature.

The map tab is only rendered at present for superusers, so this
code can be pushed to production without any public effects.

Test #418
Test #419