Relink survey data files
This migration requires you have a copy of AfESG's S3 data spool in
/u/s3 with each bucket enumerated below. If you don't care about
relinking these files, you can just make a stub directory at /u/s3
to pass this migration.

When a corresponding ID and filename are found in the S3 spool, the
file is reattached and relinked to the PopulationSubmissionAttachment
exactly as if it had been resubmitted. This creates a paper_trail
entry for the change and makes a copy of the file in the appropriate
working directory.

Unless there are exact ID and filename matches in the S3 spool, this
migration should not clobber or evict any other attachements.

Test #455