Reload all 2007 geometries
This walks all 2007 geometries and attempts to reload them from
original source -- a fresh pgsql import of the ArcGIS files from
Julian's old computer. The match is done now on

iso_code
completion_year
survey_zone
designate
short_citation

which should eliminate most opportunities to load the wrong
thing. Because it reloads all 2007 geometries, this will also
catch cases where a data manager previously linked the wrong
thing -- in the last attempt, we were leaving those alone.

Geometries which can NOT be matched on those parameters to what
was published in 2007 are unlinked from anything, to avoid the
chance of being linked to the WRONG thing which is confusing
and wrong.

One known spot case where this occurs is Tanzania's 2006
Katavi-Rukwa (Outside).

This commit clears the examples in the ticket: Boumba-Bek,
Katavi, Ruaha-Rungwa, and Gonarezhou.

Test #457