Test a fullscreen map layout
This can be tried by putting ?fullscreen=true at the end of the
replacement manager URL. It's the same functionality, with some
wrapper HTML and CSS trickery to make the interface fullscreen and
offer a left/right instead of top/bottom split. It's a 55/45
split which on an HD resolution screen (1920x1080) will yield
a usable table on the left without font size overrides, and a map
with something very close to an A4 page aspect ratio.

If smaller fonts are desired in the table, browser font overrides
work -- on Chrome these are Cmd-minus and Cmd-plus (Mac), or
Ctrl-minus and Ctrl-plus (PC).

There are plenty of nuances to discuss about this before making
it the default, but this is at least at a point worth testing. It
would be most efficient to not maintain both versions going
forward so we should try to make a decision ASAP.

Test #460