Two-phase approval system
This adds a fourth state, "Completed," to each replacement which
can only be accessed after the replacement has been put into the
"Reviewed" state.

There aren't any enforced checks around requiring separate AED
user accounts to do these two steps. That could get in the way
at some point if a single AED user account is operating the system
on behalf of multiple participants.

We can revisit enforcing the check if it seems warranted, to
guarantee that the system is not abused.

This is currently only available on the *fullscreen* replacement
manager version -- assuming here this is the one we go forward
with longer term.

Test #445