Be careful what you wish for, lest ye may receive it
This enables cascade deletes throughout the whole submission
system. I've tested it at the various deletion levels and it seems
to do what it says on the tin. paper_trail keeps a record of the
deletion and all the evicted objects.

Incidental to this, I fixed a regression from the Rails 4
migration in which the delete icons lost their "Are you sure?"
confirmations -- these now have to be in a data attribute. At
least with these restored, an ill-targeted click won't wipe out
a whole submission in an ugly and permanent way.

A behavior to examine further is what to do with the replacement
manager if a referenced stratum or submission is deleted -- this
currently results in a warning in the cell and the cell being
effectively unusable.

One possibility would be to validate and ban deletes of strata
or their parent submissions if they are used in any analysis.

Or, if we want to allow the deletes regardless, need some way
of fixing up those invalid cells.

Test #261