Generic API support for history interrogation
Hitting /history/:item_type/:item_id returns a JSON object with
the full diff history of that specific object, with the actor
(paper_trail 'whodunnit' field) resolved to an email address.

Call is reasonably efficient, but if it must be called for many
or nested objects, it could be improved.