Tooling for versioning and commenting
For discussion and testing. This is a generic Javascript-based
facility that provides access to the versions table kept by
paper_trail, and exposed usefully in the fullscreen version of
the Replacement Manager.

A pop-out window to the right of the screen shows a digested and
hopefully somewhat useful history of the linked object. In the
Replacement Manager, this opens the history of the change.

Generally, the information in the versions table is deduped and
turned into something like the kind of stream one would see in
a Skype or Slack chat, with events interspersed with comments.

Any field named "comments" is treated specially and is pulled
out in a larger typeface so that the comment history is
especially easy to follow.

Erasing the value of "comments" is also treated specially and
produces a "Resolved" notation.

Includes some minor bugfixes to the Replacement Manager for more
seamless navigation between comment histories.

Test #447
Test #450
Test #451
See #448
See #449