Replacement manager tweaks
This set of tweaks ensures that only one field is activated for
editing at any time and that any previously activated field is
committed exactly and only once. This also fixes some erroneous
flagging of comments as "Resolved" when they aren't. Tweaks the
text size of the replacement name so that field doesn't overlap
others when editing it.