Add "population" to changes table and views
This provides a place in the database to store this information
and the ability to manipulate it through the CRUD forms (old
replacement manager) at /changes. Since the Replacement Manager
is just a prettier interface on top of this CRUD, the population
can now be manipulated by an improved Replacement Manager.

See #500