Subquery corrections for ADD tables.
While this shouldn't affect any resulting data or numbers, the
subqueries for ADD did not have straightforward data, and
required the sums table to do the final math for some fields
and categories. Changed this to make the initial table a bit
more self-sufficient, so the numbers can be used in other
contexts, such as interpretations of change.

See #550