Interpretation of Change for ADD preview reports.
For the preview report, display the results of causes of
change query. This includes a migration to create a relation
table to access the display names for these causes in a way
that is friendly to stick a model in front of later for
in-application management (not supplied with this commit).

See #550