Recategorize ADD surveys.
Change surveys of type A to H, I, M and N, and change surveys
of type B to J, K and L, based on their respective estimate
types.

This commit also creates a new calculator set with the
following modifications:

* Move ADD Estimate Views creation to 1100 group
* Change estimate_locator_with_geometry_add to use ADD views

These changes ensure that the ADD categories are declared
before the metrics tables are generated.

See #525