Expand reach of ADD IoC query.
Modify the query to check for cases where there is no old
comparison data and supply zeroes as defaults (i.e. when
the reason for change is NP).

Also, catch cases where there is no new comparison data yet
by directly querying the changes table, so that new information
that is included in the changes table can be analyzed.

See #550