Analysis-centric reports, IoC stats, recalc.
Reintroduced the spatial range and area queries to the
calculator and added means of regenerating the surveys
table. Also, changed the generation logic for the IoC
ADD table to be more consistent with that of DPPS, for
comparative reasons.

Also, removed report-based mentions of 2013_africa_final
in favor of joining to the analyses table and filtering
based on the request. Changed views and queries to support
this where appropriate.

Finally, updated a couple sum counts along the way for ADD
and added a better correction for area-rendering issues in
the main statistical table.

See #550