More accurate cause of change numbers, totals.
Updated the logic behind the cause of change queries to
provide more accurate numbers across the board, as well
as more accurate totals, particular for CL.

Also, made static tables from views to make queries faster,
and added these to the /recalc command.

Finally, changed the confidence column to always be positive
based on the sum of the variance instead of the difference,
and adjusted the report display accordingly.

See #550
See #557