Appendix 2 for review
This is producing sort of the right thing, and the 2013 numbers
seem sane, but the 2007 numbers are bonkers and I can't put my
finger on why.

To access Appendix 2:
http://localhost:3000/preview_report/appendix_2/aesr_2016

Ex. 2007 Zakouma: 77,700 here. Bonkers. There aren't that many
elephants in all of Chad.

http://localhost:3000/preview_report/aesr_2016/2007/Africa/Central_Africa/Chad
shows 3,885 for Zakouma, which sounds right and also matches the
2007 AESR PDF.