Changed calculation method for other dung count.
For category C, the guess max calculation now incorporates
the confidence level, and also uses the given min/max values
instead of the estimate value, when available. This change
required a bit of a ground-up change in the way ADD values are
stored, calculated, and queried.

New calculator includes changes to the main ADD estimate table
to include appropriate column information, summary and totals
tables for each hierarchical level (country, region, continent)
and a revamp of the cause of changes tables to properly incorporate
the new guesses column.

The queries in the ADD helper class has been updated to use the
new tables, and moved the country stats query to the helper
class.

Important to note is that estimates_...reasons_for_add is gone,
which directly affects the new Appendix II calculations. This
query has been replaced by one that is compatible with the new
ADD cause of change layout.

Also, optimized cause of change queries to build a bit faster
and fixed an issue with modeled extrapolations where some could
be miscategorized.

Finally, fixed an issue with sql reader where certain comments
could cause a script to fail.

See #556
See #550
See #527