Recalculate DD reasons for ADD cause of change.
Auto-degrade reason for change that are 10+ years old and
are of reason_change "-" or "NC", instead of just "-".

Changed the calculator to have separate destroy and create
files for the base level of these queries, and added migration
to do a simple replacement of the current cause of change base
query for convenience.

A recalc will be required to see results live.