HACK: coerce range percentages
This inserts display code to coerce values >=98% to 100%, and
values <0% to 0%. This addresses some cases where small errors
between the survey and range layers, which are supposed to align
exactly, are causing numbers that are obviously incorrect.

The correct solution is to fix the layers, but this hack is
used to get a result that is intendend and publishable.